[orchagent] Move FLR timestamp to global scope
The FEC_FLR_TIMESTAMP_last should be recorded and used
in the global scope for all parts instead of having a
timestamp for each and every port in its specific RATES
table.

The timestamp should be checked before iterating over
ports to do flr calculations and if TIME_INTERVAL has
elapsed, we procedd with calculations.

Once the calculations for all ports are over, we update
the FEC_FLR_TIMESTAMP_last with the current time at the
end of all the calcualtions.

Signed-off-by: Apoorv Sachan <apoorv@arista.com>